requirements.txt
openai>=1.0.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
streamlit>=1.32.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
